lacibotframework/requirements.txt
botbuilder-integration-aiohttp>=4.14.0
openai==0.28.1
langchain==0.0.320
pysqlite3-binary; sys_platform == 'linux'
chromadb==0.4.14
tiktoken==0.5.1
